apply changes for PR(#3789) comments
