+++ modules/integration/tests-common/admin-clients/pom.xml
@@ -227,6 +227,10 @@
             <groupId>org.wso2.carbon.registry</groupId>
             <artifactId>org.wso2.carbon.registry.properties.stub</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.wso2.carbon.commons</groupId>
+            <artifactId>org.wso2.carbon.logging.admin.stub</artifactId>
+        </dependency>
     </dependencies>
 
     <properties>
